Docker container working from arm64 development machine
The is the first iteration of changes that have yeilded a working local docker container built from a
arm64 machine. The application is funtioning as expected, however, this does not include the GBMPurify
code at present. The next step will be to iterativley add in the missing elements for this code and get it
working also.