Added GBMPurity code/data and tested the UI components
Generated a docker image with the addtional UI elements required for GBMPurity and tested that
they worked in the image (which they do). Also added in the python code and data required for
the process, however i have uncommented this out at present due to a dependecy issue relating to
the python packages which i will address next.